requirements.txt
feedparser
requests
os
json
time
python-dotenv
flask